Use helpers to fiter and transform paths in detectors
