+++ lib/crystalball/map_generator/coverage_strategy.rb
@@ -12,7 +12,7 @@ module Crystalball
 
       attr_reader :execution_detector
 
-      def initialize(execution_detector = ExecutionDetector.new(Dir.pwd))
+      def initialize(execution_detector = ExecutionDetector.new)
         @execution_detector = execution_detector
       end
